Chore #161269373 Fix travis after folder restructure
