+++ app/tests/v1/test_v1_sales.py
@@ -1,6 +1,6 @@
 import json
 
-from app.tests.base import BaseTestCase
+from app.tests.v1.base import BaseTestCase
